requirements.txt
bs4
requests
azure-functions
azure-storage-blob